refactor(model): make metacriticRating a Double across entity, DTO, and services
Align the metacritic score type with imdbRating/tmdbRating so ratings compose cleanly (e.g. fallback chains) without per-site int/double conversion.

+++ JMedia/com.playdeca.JMedia/src/main/java/Models/Video.java
@@ -59,7 +59,7 @@ public class Video extends PanacheEntity {
     // Entertainment Metadata
     public Double imdbRating = 0.0;
     public Double tmdbRating = 0.0;
-    public Integer metacriticRating;
+    public Double metacriticRating;
     @Column(length = 1000)
     public String awards;
     public Double userRating = 0.0;
